// src/app/shared/modules/interface.ts
export interface Iobject {
    fname: string;
    lname: string;
    email: string;
    contact: number;
}

export interface Heroes {
    fname: string;
    lname: string;
    email: string;
    contact: number;
}

export interface Countries {
    countryname: string;
    countrycapital: string;
    countrylanguage: string;
    countrycurrency: string;
}